Reorganize all tests to tests/ dir
