+++ tests/sanity/cinder/fakecloud.go
