+++ tests/sanity/cinder/fakemetadata.go
