+++ tests/sanity/manila/fakemanilaclient.go
